🌱 refactor: extract T2_TEMPLATES from CardFactoryModal into dedicated file
CardFactoryModal.tsx held 484 lines of static template data inline
(interface T2Template + const T2_TEMPLATES). Extracting to a dedicated
file reduces the modal from 1227 → 743 lines, keeping related logic
together and template data separately browsable.

Changes:
- New file: cardFactoryTemplatesT2.ts  (486 lines of template data)
- CardFactoryModal.tsx: 1227 → 743 lines (-484 lines)
- COPY_FEEDBACK_TIMEOUT_MS moved to T2 file (its only usage site)
- Import added alongside existing T1_TEMPLATES import

Pattern mirrors the existing T1_TEMPLATES → cardFactoryTemplates.ts
extraction from previous passes. Zero behavior change.

Part of console#8608 file-split series.
